requirements.txt
streamlit>=1.40
pandas>=2.2
openpyxl>=3.1
plotly>=5.24
google-genai>=1.0.0
